requirements.txt
tweepy==4.10.0
requests==2.28.1
ics==0.7.2
Pillow==9.2.0